move RelationAttribute to attributes package
